Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishesForRestaurant/DeleteDishesForRestaurantCommandHandler.cs
@@ -1,14 +1,17 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishesForRestaurant;
 
 public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRestaurantCommandHandler> logger,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishesForRestaurantCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantsAuthorizationService restaurantsAuthorizationService) : IRequestHandler<DeleteDishesForRestaurantCommand>
 {
     public async Task Handle(DeleteDishesForRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -16,6 +19,9 @@ public class DeleteDishesForRestaurantCommandHandler(ILogger<DeleteDishesForRest
 
         var restaurant = await restaurantsRepository.GetByIdAsync(request.RestaurantId);
         if (restaurant is null) throw new NotFoundException(nameof(Restaurant),request.RestaurantId.ToString());
+
+        if (!restaurantsAuthorizationService.Authorize(restaurant, ResourceOperation.Delete))
+            throw new ForbiddenException();
 
         await dishesRepository.Delete(restaurant.Dishes);
     }
