+++ Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -1,14 +1,16 @@
-﻿using AutoMapper;
-using MediatR;
+﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
 public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger,
-    IRestaurantsRepository restaurantsRepository) : IRequestHandler<DeleteRestaurantCommand>
+    IRestaurantsRepository restaurantsRepository,
+    IRestaurantsAuthorizationService restaurantsAuthorizationService) : IRequestHandler<DeleteRestaurantCommand>
 {
     public async Task Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -18,6 +20,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
 
         if (restaurant is null)
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
+
+        if(!restaurantsAuthorizationService.Authorize(restaurant,ResourceOperation.Delete))
+            throw new ForbiddenException();
 
         await restaurantsRepository.Delete(restaurant);
     }
